+++ packages/react-core/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [6.3.0-prerelease.27](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@6.3.0-prerelease.26...@patternfly/react-core@6.3.0-prerelease.27) (2025-06-23)
+
+### Bug Fixes
+
+- Bumped to latest core version. ([#11880](https://github.com/patternfly/patternfly-react/issues/11880)) ([0380aa5](https://github.com/patternfly/patternfly-react/commit/0380aa504b4d0d83459126bfd98b2e4e29cac4e8))
+
+# [6.3.0-prerelease.26](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@6.3.0-prerelease.25...@patternfly/react-core@6.3.0-prerelease.26) (2025-06-20)
+
+### Features
+
+- **Table:** added optin animations for expansion ([#11865](https://github.com/patternfly/patternfly-react/issues/11865)) ([9adc88b](https://github.com/patternfly/patternfly-react/commit/9adc88b652f77ed5737c9f6dd02885daace1cd2b))
+
+# [6.3.0-prerelease.25](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@6.3.0-prerelease.24...@patternfly/react-core@6.3.0-prerelease.25) (2025-06-18)
+
+### Features
+
+- **SearchInput:** enable animations ([#11874](https://github.com/patternfly/patternfly-react/issues/11874)) ([50ddd0f](https://github.com/patternfly/patternfly-react/commit/50ddd0f2b8f0c0e53caf4be6a9fa4b2b15f13ceb))
+
 # [6.3.0-prerelease.24](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-core@6.3.0-prerelease.23...@patternfly/react-core@6.3.0-prerelease.24) (2025-06-17)
 
 ### Features