+++ packages/react-icons/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [6.3.0-prerelease.9](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-icons@6.3.0-prerelease.8...@patternfly/react-icons@6.3.0-prerelease.9) (2025-06-23)
+
+### Bug Fixes
+
+- Bumped to latest core version. ([#11880](https://github.com/patternfly/patternfly-react/issues/11880)) ([0380aa5](https://github.com/patternfly/patternfly-react/commit/0380aa504b4d0d83459126bfd98b2e4e29cac4e8))
+
+# [6.3.0-prerelease.8](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-icons@6.3.0-prerelease.7...@patternfly/react-icons@6.3.0-prerelease.8) (2025-06-20)
+
+### Features
+
+- **Table:** added optin animations for expansion ([#11865](https://github.com/patternfly/patternfly-react/issues/11865)) ([9adc88b](https://github.com/patternfly/patternfly-react/commit/9adc88b652f77ed5737c9f6dd02885daace1cd2b))
+
 # [6.3.0-prerelease.7](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-icons@6.3.0-prerelease.6...@patternfly/react-icons@6.3.0-prerelease.7) (2025-06-17)
 
 ### Bug Fixes