+++ packages/react-table/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [6.3.0-prerelease.27](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@6.3.0-prerelease.26...@patternfly/react-table@6.3.0-prerelease.27) (2025-06-23)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [6.3.0-prerelease.26](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@6.3.0-prerelease.25...@patternfly/react-table@6.3.0-prerelease.26) (2025-06-20)
+
+### Features
+
+- **Table:** added optin animations for expansion ([#11865](https://github.com/patternfly/patternfly-react/issues/11865)) ([9adc88b](https://github.com/patternfly/patternfly-react/commit/9adc88b652f77ed5737c9f6dd02885daace1cd2b))
+
+# [6.3.0-prerelease.25](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@6.3.0-prerelease.24...@patternfly/react-table@6.3.0-prerelease.25) (2025-06-18)
+
+**Note:** Version bump only for package @patternfly/react-table
+
 # [6.3.0-prerelease.24](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@6.3.0-prerelease.23...@patternfly/react-table@6.3.0-prerelease.24) (2025-06-17)
 
 **Note:** Version bump only for package @patternfly/react-table